requirements.txt
# Example dependencies
requests
datetime
python-dateutil